Refactor to avoid tsconfig error
